Prevent brute force attack #19

+++ knights/table/outchain/random_val.hpp
@@ -1,10 +1,10 @@
 #pragma once
 
 struct random_val {
-    uint64_t seed;
+    uint32_t seed;
     uint32_t value;
 
-    random_val(uint64_t seed, uint32_t value) {
+    random_val(uint32_t seed, uint32_t value) {
         this->seed = seed;
         this->value = value;
     }
